Read me
This is a simple image downloader, you can use this program to download images from the internet, you only need to paste the url to the full path of the image to download it. If you paste the wrong url the program will not download the image properly. More functionality to be added later. 

Thanks for visiting my github page...