Image downloader
I am Francis F. Massaquoi, Jr. a beginner in python, this is my first python program it is an image downloader, I created it using python, you need to paste the full url of the image you want to download in the program.

Note: 
If this program don't work for python2, try python 3.